+++ tests/Js/Main/HashSetTests.fs
@@ -6,13 +6,18 @@ open Util.Testing
 type MyRefType(i: int) =
     member x.Value = i
 
+type MyRefTypeComparer() =
+    interface IEqualityComparer<MyRefType> with
+        member _.Equals(x, y) = x.Value = y.Value
+        member _.GetHashCode(x) = x.Value
+
 type IgnoreCaseComparer() =
     interface IEqualityComparer<string> with
         member _.Equals(s1, s2) =
-            s1.Equals(s2, System.StringComparison.InvariantCultureIgnoreCase)
+            System.String.Equals(s1, s2, System.StringComparison.InvariantCultureIgnoreCase)
 
         member _.GetHashCode(s) =
-            s.ToLowerInvariant().GetHashCode()
+            if System.String.IsNullOrEmpty(s) then 0 else s.ToLowerInvariant().GetHashCode()
 
 let inline hashSet l =
     let xs = HashSet<_>()
@@ -55,11 +60,7 @@ let tests =
         set1.Contains(x) |> equal true
         set1.Contains(y) |> equal false
 
-        let comparer =
-            { new IEqualityComparer<MyRefType> with
-                member _.Equals(x, y) = x.Value = y.Value
-                member _.GetHashCode(x) = x.Value }
-        let set2 = HashSet<_>(comparer)
+        let set2 = HashSet<_>(MyRefTypeComparer())
         set2.Add(x) |> equal true
         set2.Contains(x) |> equal true
         set2.Contains(y) |> equal true
@@ -162,10 +163,6 @@ let tests =
         xs.SetEquals ["foo"; "BAR"; "foo"] |> equal true
         xs.SetEquals ["foo"; "baz"] |> equal false
 
-    testCase "HashSet creation works" <| fun () ->
-        let hs = HashSet<_>()
-        equal 0 hs.Count
-
     testCase "HashSet iteration works" <| fun () ->
         let hs = HashSet<_>()
         for i in 1. .. 10. do hs.Add(i*i) |> ignore
@@ -203,7 +200,7 @@ let tests =
         hs.Add(3) |> equal true
         hs.Count |> equal 2
 
-    testCase "HashSet.Add works ||" <| fun () ->
+    testCase "HashSet.Add works II" <| fun () ->
         let hs = HashSet<_>()
         hs.Add("A") |> equal true
         hs.Add("B") |> equal true
@@ -293,6 +290,33 @@ let tests =
         xs.IsProperSupersetOf(ys) |> equal true
         ys.IsProperSupersetOf(xs) |> equal false
         xs.IsProperSupersetOf(xs) |> equal false
+
+    testCase "HashSet.CopyTo works" <| fun () ->
+        let xs = HashSet<_>()
+        for i in 1 .. 9 do xs.Add(i) |> ignore
+
+        let arr1 = Array.zeroCreate 9
+        let arr2 = Array.zeroCreate 11
+        let arr3 = Array.zeroCreate 7
+
+        xs.CopyTo(arr1)         // [|x;x;x;x;x;x;x;x;x|]
+        xs.CopyTo(arr2, 2)      // [|0;0;x;x;x;x;x;x;x;x;x|]
+        xs.CopyTo(arr3, 3, 4)   // [|0;0;0;x;x;x;x|]
+
+        let sum = fun acc item -> acc + item
+        arr1 |> Seq.fold sum 0 |> equal 45
+        arr1.Length |> equal 9
+
+        arr2 |> Seq.fold sum 0 |> equal 45
+        arr2.Length |> equal 11
+
+        // The exact order of the copied items is not guaranteed,
+        // but they must all exist in the hash set and be distinct.
+        // arr3 |> Seq.fold sum 0 |> equal 10 // not guaranteed to work
+        arr3[0..2] |> Array.forall ((=) 0) |> equal true
+        arr3[3..] |> Array.forall xs.Contains |> equal true
+        arr3[3..] |> Array.distinct |> Array.length |> equal 4
+        arr3.Length |> equal 7
 
     testCase "HashSet IReadOnlyCollection.Count works" <| fun _ ->
         let xs = [| ("A", 1); ("B", 2); ("C", 3) |]